Made some cahnges for validation and implementing heroku

+++ controllers/taskController.js
@@ -3,6 +3,7 @@ const path = require("path");
 const Task = require("../models/taskSchema");
 const uniqid = require("uniqid");
 const dotenv = require("dotenv");
+const { userInfo } = require("os");
 dotenv.config({ path: "./config.env" });
 const fileName = path.join(__dirname, "..", "data", "listData.json");
 const taskData = JSON.parse(fs.readFileSync(fileName, "utf-8"));
@@ -39,6 +40,11 @@ const getTaskById = (req, res) => {
 };
 
 const createTask = async(req, res) => {
+    if (!req.body.taskName) {
+        return res.status(400).send({
+            message: "The Task Name cannote be empty.",
+        });
+    }
     try {
         const data = new Task({
             taskId: uniqid(),
@@ -59,6 +65,11 @@ const createTask = async(req, res) => {
 };
 
 const deleteTask = async(req, res) => {
+    if (!req.params.taskId) {
+        return res.status(404).send({
+            message: "User not found",
+        });
+    }
     try {
         const id = req.params.taskId;
         let result = await Task.deleteOne({ taskId: id });
@@ -75,11 +86,16 @@ const deleteTask = async(req, res) => {
 };
 
 const updateTask = async(req, res) => {
+    if (!req.body.taskName) {
+        return res.status(400).send({
+            message: "The Task Name cannote be empty.",
+        });
+    }
     try {
         const status = req.params.status;
         let result = await Task.findOneAndUpdate(status, req.body);
         res.status(200).json({
-            message: "Updated Sucesfully",
+            message: "Updated Sucessfully",
             data: result,
         });
     } catch (err) {
